+++ android/src/main/java/com/dooboolab/RNIap/DoobooUtils.java
@@ -1,5 +1,6 @@
 package com.dooboolab.RNIap;
 
+import android.os.Build;
 import android.util.Log;
 
 
@@ -42,8 +43,8 @@ public class DoobooUtils {
   public static final String APPSTORE_GOOGLE = "GOOGLE_PLAY";
   public static final String APPSTORE_AMAZON = "AMAZON";
 
-  private HashMap<String, ArrayList<Promise>> promises = new HashMap<>();
-  private static DoobooUtils instance = new DoobooUtils();
+  private final HashMap<String, ArrayList<Promise>> promises = new HashMap<>();
+  private static final DoobooUtils instance = new DoobooUtils();
 
   public static DoobooUtils getInstance() {
     return instance;
@@ -56,7 +57,7 @@ public class DoobooUtils {
         list = promises.get(key);
       }
       else {
-        list = new ArrayList<Promise>();
+        list = new ArrayList<>();
         promises.put(key, list);
       }
 
@@ -201,20 +202,35 @@ public class DoobooUtils {
     return array;
   }
 
+    private static final String AMAZON_FEATURE_FIRE_TV = "amazon.hardware.fire_tv";
+  private static final String AMAZON_FIRE_TV_MODEL_PREFIX = "AFT";
+
+  /**
+   * Detects Stores:
+   * Amazon tablets and Fire TV are considered as APPSTORE_AMAZON
+   * using recommended detection logic from:
+   * https://developer.amazon.com/docs/fire-tv/identify-amazon-fire-tv-devices.html
+   * @param context Application context
+   * @return One of APPSTORE_AMAZON,APPSTORE_GOOGLE,APPSTORE_UNKNOWN
+   *
+   * Suppressing deprecation since the alternative requires API level 30
+   */
+  @SuppressWarnings("deprecation")
   public final String getInstallSource(Context context) {
     Context appContext = context.getApplicationContext();
     PackageManager pkgManager = appContext.getPackageManager();
     String installerPackageName = pkgManager.getInstallerPackageName(appContext.getPackageName());
 
-    if (installerPackageName == null) {
-      return APPSTORE_UNKNOWN;
+    if (pkgManager.hasSystemFeature(AMAZON_FEATURE_FIRE_TV) ||
+            (installerPackageName!=null && installerPackageName.startsWith("com.amazon.")) ||
+            Build.MODEL.startsWith(AMAZON_FIRE_TV_MODEL_PREFIX)) {
+      Log.d(TAG, "Yes, this is a Fire TV device.");
+      return APPSTORE_AMAZON;
     } else if ("com.android.vending".equals(installerPackageName)) {
       return APPSTORE_GOOGLE;
-    } else if (installerPackageName.startsWith("com.amazon.")) {
-      return APPSTORE_AMAZON;
-    } else {
+    }  else {
       Log.d(TAG, "Unknown installer source: " + installerPackageName);
+      return APPSTORE_UNKNOWN;
     }
-    return APPSTORE_UNKNOWN;
   }
 }